Fix/filesize

+++ packages/itmat-commons/src/graphql/files.ts
@@ -1,7 +1,7 @@
 import gql from 'graphql-tag';
 
 export const UPLOAD_FILE = gql`
-    mutation uploadFile($studyId: String!, $file: Upload!, $description: String!, $fileLength: Int, $hash: String) {
+    mutation uploadFile($studyId: String!, $file: Upload!, $description: String!, $fileLength: BigInt, $hash: String) {
         uploadFile(studyId: $studyId, description: $description, file: $file, fileLength: $fileLength, hash: $hash) {
             id
             fileName
